requirements.txt
pytesseract
pymupdf
pillow
numpy
